MARWEB-48 Display AIS data in a smooth lightweight manner. Using the new dynamic webservice which retrieves ais vessels in a specific  area. That mean that vessels are only displayed within the bounding box that is visible, and thus relevant, to them. if you sail between Helsinki and Talin, you won't waste bandwith on fishing boats moving back and forth Bornholm.
Using the new data lightweight  overview webservice (from MARWEB-54) and drawing simplified vessels when zoom level is above 8. In the overview mode vessels are displayed with a small circle. The solution is lighter in data usage, but does not appear to significantly better performance-vise when displaying more than 3K vessels.